server/requirements.txt
Django==1.4.3
South==0.7.6
django-extensions==1.0.3
django-tastypie==0.9.11
gunicorn==0.17.2
lxml==3.1beta1
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==1.5
wsgiref==0.1.2
xmltodict==0.8.3
